Make int->float rounding explicit to pacify clang's -Wimplicit-int-float-conversion
(cherry picked from commit 1ab227ccb936d7f22c8062be3cad6b330920d69f)

+++ src/common/opencl.c
@@ -928,7 +928,7 @@ static void encrypt_tea(unsigned int *arg)
 
 static float tpdf(unsigned int urandom)
 {
-  float frandom = (float)urandom / 0xFFFFFFFFu;
+  float frandom = (float)urandom / (float)0xFFFFFFFFu;
 
   return (frandom < 0.5f ? (sqrtf(2.0f * frandom) - 1.0f) : (1.0f - sqrtf(2.0f * (1.0f - frandom))));
 }
